separate sla-class-list
