+++ catalystwan/endpoints/configuration/policy/list/sla.py
@@ -5,17 +5,9 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import SLAClassList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.sla import SLAClassList, SLAClassListEditPayload, SLAClassListInfo
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class SLAClassListEditPayload(SLAClassList, PolicyListId):
-    pass
-
-
-class SLAClassListInfo(SLAClassList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicySLAClassList(APIEndpoints, PolicyListEndpoints):
